First version of documentation

+++ server/app/Http/Controllers/API/CategoryController.php
@@ -12,19 +12,14 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
 /**
- * @tags Categories
+ * @tags Client categories
  */
 
 class CategoryController extends Controller
 {
 
-    public function __construct(){
-        $this->middleware('auth:sanctum')->only(["destroy"]);
-    }
-
     /**
-     * Display every clients categories
-     * @unauthenticated
+     * Display all client categories
      */
 
     public function index(Request $request)
@@ -41,7 +36,7 @@ class CategoryController extends Controller
         }
         return 
         /**
-         * Returns every clients categories in the database. The clients tokens will be displayed if you are authenticated.
+         * Returns all client categories in the database. The clients tokens will be displayed if you are authenticated.
          * @body array{array{id: int, name: string, client_id: int, client: array{id: int, name: string, token: string|null}}}
          */
         $categories;
@@ -104,7 +99,7 @@ class CategoryController extends Controller
                 } else {
                     return 
                     /**
-                     * Error thrown when the image link leads to an incorrect extension.
+                     * Error thrown when one of the image links leads to an incorrect extension.
                      * @status 415
                      */
                     response()->json(["message"=>"The image link is not in png or jpg"],415);
@@ -116,7 +111,7 @@ class CategoryController extends Controller
 
         return 
         /**
-         * Returns the client category with its brands.
+         * Returns the new client's category with its brands.
          * @body array{id: int, name: string, client_id: int, client: array{id: int, name: string, token: string}, brands: array{array{id: int, png_url: string|null, jpg_url: string|null, validated: bool}}}
          */
         response()->json($category->load("client","brands"),201);
@@ -125,7 +120,6 @@ class CategoryController extends Controller
 
     /**
      * Find a specific client category.
-     * @unauthenticated
      */
     public function show(Category $category)
     {
@@ -148,23 +142,42 @@ class CategoryController extends Controller
     }
 
     /**
-     * Update the specified resource in storage.
+     * Update a specific client category data.
+     * 
+     * The updated category need to have **at least one brand** linked to it. \
+     * For each brands, the back-end process will check if the name already exists in the database :
+     * * If the case written above is **true** and the brand **is not already associated to the category**, the already existing entry is linked to the category.
+     * * If the case written above is **false**, a new brand is inserted in the database, but its validation state is set to false. **Only an authenticated user can change the validation state.** The newly created brand will be linked to the category.
+     * If a brand previously associated to the category is not in the body data, it will be automatically dissociated from the category.
      * @unauthenticated
      */
     public function update(Request $request, Category $category)
     {
         $request->validate([
+            //The client id, it must be the client that owns the category
             "client_id"=>["required","integer","exists:clients,id", new ClientOwnsCategory($category)],
+            //The client token
             "client_token"=>"required|string|exists:clients,token",
+            //The category name, it must be different than the others client's categories.
             "name"=>["required","string",new CheckCategoryName(Client::findOrFail($request->input("client_id") ? $request->input("client_id") : 1), $category)],
             "brands"=>"required|array",
             "brands.*"=>"required|array",
+            //The brand name. The value will be parsed into the correct naming format.
             "brands.*.name"=>"required|string",
+            /**
+             * The image url, it has to be a png or jpg file.
+             * @example https://upload.wikimedia.org/wikipedia/commons/4/47/PNG_transparency_demonstration_1.png
+             */
             "brands.*.url"=>"nullable|url"
         ]);
 
         if(Client::find($request->input("client_id"))->token != $request->input("client_token")){
-            return response()->json(["message"=>"Client id and token are not matching"],401);
+            return 
+             /**
+             * Error thrown when a wrong id/token combination is given.
+             * @status 401
+             */
+            response()->json(["message"=>"Client id and token are not matching"],401);
         }
 
         $inputBrandNames = array_map(function($v){
@@ -190,7 +203,12 @@ class CategoryController extends Controller
                     } else if(str_starts_with($extension,"jpg") || str_starts_with($extension,"jpeg")){
                         $brandData["jpg_url"] = $brand["url"];
                     } else {
-                        return response()->json(["message"=>"The image link is not in png or jpg"],500);
+                        return 
+                         /**
+                         * Error thrown when one of the image links leads to an incorrect extension.
+                         * @status 415
+                         */
+                        response()->json(["message"=>"The image link is not in png or jpg"],500);
                     }
                     $brandModel = Brand::create($brandData);
                 }
@@ -198,15 +216,37 @@ class CategoryController extends Controller
             }
         }
 
-        return response()->json($category->load("brands"));
+        return 
+        /**
+         * Returns the updated category with its brands.
+         * @body array{id: int, name: string, client_id: int, client: array{id: int, name: string, token: string}, brands: array{array{id: int, png_url: string|null, jpg_url: string|null, validated: bool}}}
+         */
+        response()->json($category->load("brands"));
     
     }
 
     /**
-     * Remove the specified resource from storage.
+     * Delete a specific client category from the database.
+     * 
+     * **WARNING :** Only the category and the brands associations will be deleted. You must use a different method to directly delete brands.
+     * @unauthenticated
      */
-    public function destroy(Category $category)
+    public function destroy(Category $category, Request $request)
     {
+        $request->validate([
+            //The client id, it must be the client that owns the category
+            "client_id"=>["required","integer","exists:clients,id", new ClientOwnsCategory($category)],
+            //The client token
+            "client_token"=>"required|string|exists:clients,token"
+        ]);
+        if(Client::find($request->input("client_id"))->token != $request->input("client_token")){
+            return 
+             /**
+             * Error thrown when a wrong id/token combination is given.
+             * @status 401
+             */
+            response()->json(["message"=>"Client id and token are not matching"],401);
+        }
         $category->brands()->detach();
         $category->delete();
         return response("",204);
